fix: remove quirrel and switch to croner
