+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.8.0",
   "private": true,
   "scripts": {
-    "dev": "concurrently -c bgGreen,bgBlue -n vite,quirrel \"vite dev --port 5173\" \"quirrel\"",
-    "build": "vite build && quirrel ci",
+    "dev": "vite dev --port 5173",
+    "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
@@ -53,6 +53,7 @@
     "blurhash": "^2.0.5",
     "clsx": "^2.1.0",
     "cmdk-sv": "^0.0.17",
+    "croner": "^8.0.2",
     "dompurify": "^3.1.0",
     "embla-carousel-svelte": "^8.0.4",
     "formsnap": "^1.0.0",
@@ -61,7 +62,6 @@
     "marked": "^12.0.1",
     "minimatch": "^9.0.4",
     "mode-watcher": "^0.3.0",
-    "quirrel": "^1.14.1",
     "svelte-sonner": "^0.3.22",
     "sveltekit-superforms": "^2.12.4",
     "tailwind-merge": "^2.2.2",
